Add files via upload
I believe that the reference for the file scanner I had to change for the file I chose. So if you want to switch text files you might need to change its pointer. Also it has a file directory so maybe just create a file in your C drive and then adjust its pointer.